// frontend-ui/src/app/services/models/main-balance-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MainBalanceResponseDto {
  earnings?: number;
  id?: number;
  lastUpdated?: string;
  message?: string;
  totalBalance?: number;
  totalInstallmentReturn?: number;
  totalInvestment?: number;
  totalMaintenanceCost?: number;
  totalProductCost?: number;
  totalWithdrawal?: number;
}